==> notebooks/step02_clean_combined_df.ipynb <==
## Healthcare Bluebook EDA #2

### Updated 6/2/2020:
- Moved the code that eliminated the unnecessary row from this notebook. Put it in the "step01..." notebook, so that the pickle file won't have that row in it.

## Step 2: Cleaning the combined df based on discoveries made during initial EDA.
1. Eliminate irrelevant row from 2015 (similar to a footnote, not needed)
2. Add column to show payment_type (Doctor Only, Facility Only, Doctor & Facility)
3. Create df_payments_combined file to use for the rest of the process.


### See next notebook named "step03_EDA_and_analysis" for additional EDA and work torwards our deliverables

### Drop irrelevant row from 2015

import pandas as pd
import pickle
import numpy as np

df_payments_combined_draft = pd.read_pickle('../data/pickled_files/payments_combined_draft.pkl')
print(df_payments_combined_draft.shape)
df_payments_combined_draft.head()

### Adding column to indicate whether payment is Doctor Only, Facility Only, or Doctor & Facility

**Entity Type**  
I = Individual  
O = Organization  

**Place of Service**  
O = Office  
F = Facility  

**Entity, Place = Description**  
I, F = Doctor Only  
I, O = Doctor & Facility
O, F = Facility Only  
O, O = Doctor & Facility


# Step 1 to add new column with payment type (Code from Deigo)

conditions = [
    (df_payments_combined_draft.place_of_service == 'O'),
    (df_payments_combined_draft.entity_type_of_the_provider == 'I') & (df_payments_combined_draft.place_of_service == 'F'),
    (df_payments_combined_draft.entity_type_of_the_provider == 'O') & (df_payments_combined_draft.place_of_service == 'F')
]

choices = ['Doctor & Facility', 'Doctor Only', 'Facility Only']

# Step 2 to add new column with payment type (Code from Deigo)

df_payments_combined_draft['payment_type'] = np.select(conditions, choices, default = 'unknown')

# Confirmed that there are no "unknown" values.

df_payments_combined_draft['payment_type'].unique()

# Double checking that it work. All set!

df_payments_combined_draft.tail()

### Rename df to the final name the team is using:  df_payments_combined

# To rename df to match the class, after cleaning

df_payments_combined = df_payments_combined_draft
df_payments_combined.head()

df_payments_combined.to_pickle('..\data\pickled_files\payments_combined.pkl')

### See next notebook named "step03_EDA_and_analysis" for additional EDA and work torwards our deliverables
